\name{Balt.ts}
\alias{Balt.ts}
\docType{data}
\title{Baltimore Electricity Load and Temperature Data}
\description{
timeSeries object
}
\usage{data(Balt.ts)}
\details{
 Details are given in the statement of the Problem concerned with the analysis of this data set
}
\source{R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag}
\author{Rene Carmona, \email{rcarmona@princeton.edu}}
\examples{
data(Balt.ts)plot(Balt.ts[,1])
}
\keyword{datasets}
