\name{BASKETBALL}
\alias{BASKETBALL}
\docType{data}
\title{Basket ball data used in regression problems}
\usage{data(BASKETBALL)}
\format{
  A data frame with 96 observations on the following 6 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
    \item{\code{V3}}{a numeric vector}
    \item{\code{V4}}{a numeric vector}
    \item{\code{V5}}{a numeric vector}
    \item{\code{V6}}{a factor with levels \code{A.Bowie} \code{A.Johnson} \code{A.Peeler} \code{A.Robertson} \code{B.Coles} \code{B.J.Armstrong} \code{B.Price} \code{B.Shaw} \code{B.Smith} \code{C.Drexler} \code{C.Smith} \code{D.Ainge} \code{D.Barros} \code{D.Brown} \code{D.Cooper} \code{D.Harper} \code{D.Overton} \code{D.Rivers} \code{D.Walker} \code{D.West} \code{D.Young} \code{E.Murdock} \code{F.Johnson} \code{G.Anthony} \code{G.Grant} \code{G.Payton} \code{G.Wilkins} \code{H.Davis} \code{H.Hawkins} \code{H.Miner} \code{I.Thomas} \code{J.Barry} \code{J.Battle} \code{J.Dawkins} \code{J.Dumars} \code{J.Hamphries} \code{J.Hornacek} \code{J.Jackson} \code{J.Les} \code{J.Malone} \code{J.Paxson} \code{J.Starks} \code{J.Stockton} \code{K.Anderson} \code{K.Edwards} \code{K.Gill} \code{K.Jennings} \code{K.Johnson} \code{K.Smith} \code{L.Daniels} \code{L.Green} \code{L.Mayberry} \code{L.Smith} \code{L.Sprewell} \code{M.Abdul-Rauf} \code{M.Adams} \code{M.Blaylock} \code{M.Cheeks} \code{M.Jackson} \code{M.Jordan} \code{M.Macon} \code{M.Price} \code{M.Richmond} \code{M.Wiley} \code{M.Williams} \code{N.Knight} \code{P.Richardson} \code{R.Blackman} \code{R.Brown} \code{R.Chapman} \code{R.Harper} \code{R.Miller} \code{R.Pack} \code{R.Pierce} \code{R.Robinson} \code{R.Strickland} \code{S.Brooks} \code{S.Douglas} \code{S.Floyd} \code{S.Henson} \code{S.Marciulionis} \code{S.Skiles} \code{S.Smith} \code{S.Threatt} \code{S.Webb} \code{T.Bennett} \code{T.Bogues} \code{T.Brandon} \code{T.Hardaway} \code{T.Legler} \code{T.Porter} \code{T.Smith} \code{T.Tucker} \code{V.D.Negro} \code{V.Fleming} \code{V.Maxwell}}
  }
}
\details{
      V1     V2  V3    V4 V5         V6
1 0.5885 0.0888 201 36.02 28   R.Miller
2 0.8291 0.1399 198 39.32 30   M.Jordan
3 0.4974 0.0747 198 38.80 26     D.West
4 0.5772 0.0983 191 40.71 30   J.Dumars
5 0.5703 0.1276 196 38.40 28 M.Richmond
6 0.5835 0.1671 201 34.10 31  C.Drexler
> }
\references{
R. A. Carmona: Statistical Analysis of Financial Data in S-Plus, (2004) Springer Verlag, Chapter 3}
\examples{
data(BASKETBALL)
## maybe str(BASKETBALL) ; plot(BASKETBALL) ...
}
\keyword{datasets}
