\name{BCofLRet}
\alias{BCofLRet}
\alias{CCofLRet}
\docType{data}
\title{Coffee log-returns}
\description{
  Numeric vector giving the daily log-returns on the close of the Front Month futures contract of Brazilian coffee (\code{BCofLRet}) and Colombian coffee (\code{CCofLRet}).
}
\usage{data(BCofLRet)}
\format{
  Object of a numeric vector.
}
\source{
R. A. Carmona: Statistical Analysis of Financial Data in S-Plus, (2004) Springer Verlag, Chapter 2 }
\examples{
data(BCofLRet)
data(CCofLRet)
## maybe str(BCofLRet) ; plot(BCofLRet) ...
}
\keyword{datasets}
