\name{co2.ts}
\alias{co2.ts}
\docType{data}
\title{
CO2 concentration
}
\description{
Concentration of CO2 above the Mauna Loa volcano in Hawaii. The data comprisemonthly measurements. They range from March 1958 to December 2008
}
\usage{head(co2.ts)}
\format{
The data is an object of class timeSeries}
\references{
 R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
}
\examples{
plot(co2.ts)
}
\keyword{datasets}
