\name{dmvnorm}
\alias{dmvnorm}
\alias{pmvnorm}
\alias{rmvnorm}

\title{ Multivariate Normal (Gaussian) distribution }

\description{
The mean of the distribution may be a d-dimensional vector. The covariance matrix should be given as a d x d non-negative definite matrix if supplied with the parameter \code{cov}. It can also be given,  if cov is missing, by a vector \code{sd} for the marginal standard deviations and a scalar \code{rho} implying a constant correlation between all the marginals. If the covariance structure of the marginals is supplied in this way,  \code{sd} should be a d-dimensional vector,
 and \code{rho} should be scalar. The dimension \code{d} may be inferred from other arguments.
The code of this function is a mere wrapper for the function with the same name from the library \code{mvtnorm}. It was written to provide compatibility with S-Plus, hence the long list of parameters
}

\details{
\code{dmvnorm},compute multivariate normal density.
\code{pmvnorm} compute multivariate normal c.d.f.
\code{rmvnorm} generate random samples from the multivariate normal distribution.
}
\usage{
dmvnorm(x, mean = rep(0, d), cov = diag(d), sd, rho, d = 2, sigma = cov, log = FALSE)

pmvnorm(x, mean = rep(0, d), cov = diag(d), sd, rho, d = 2, sigma = cov, log = FALSE)

rmvnorm(n, mean = rep(0, d), cov = diag(d), sd, rho, d = 2, sigma = cov, log = FALSE)
}

\arguments{
  \item{n}{Number of samples generated by \code{rmvnorm}}
   \item{x}{ n x d numeric matrix, each row giving a point at which the density is computed}
  \item{mean}{ d- dimensional vector giving the mean of the distribution }
  \item{cov}{ d x d matrix giving the covariance matrix of the distribution }
  \item{sd}{ d- vector of the marginal standard deviations }
  \item{rho}{ number giving the constant correlation when the covariance matrix is given by its diagonal and the parameter \code{rho} }
  \item{d}{ dimension of the distribution }
  \item{sigma}{ used for compatibility with S-Plus }
  \item{log}{boolean for logarithmic scale }
   \item{method}{String giving the method SVD or Choleski used}
}

\value{
A list of the elements  
  \item{$x}{n x d Matrix giving the values where the density is computed}
  \item{$y}{Vector of length \code{n} giving the values of the density}
}
\references{library \code{Rmetrics} and \code{S-Plus} manual }
\author{Rene Carmona}

\seealso{\code{\link{dnorm}}}
\examples{
\dontrun{
TSAMPLE <- rmvnorm(n=128, mean=rep(0,2), sd=rep(1,2), rho=.18) 
TDENS <- kdest(TSAMPLE[,1], TSAMPLE[,2]) 
}
}
\keyword{distribution}
