\name{DSP.ts}
\alias{DSP.ts}
\docType{data}
\title{
Daily S&P close
}
\description{
timeSeries object
}
\usage{data(DSP.ts)}
\details{
 Details are given in the text
}
\source{R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag}
\author{Rene Carmona, \email{rcarmona@princeton.edu}}
\examples{
plot(DSP.ts)
}
\keyword{datasets}
