\name{DUKE.wer}
\alias{DUKE1.wer}
\alias{DUKE2.wer}
\docType{data}
\title{Weekly Excess Returns for Duke Energy}
\description{
 Numeric vector giving the weekly excess returns for Duke Energy. The vectors DUKE1.wer and DUKE2.wer are subsets of DUKE.wer covering disjoint periods discussed in the reference text. 
}
\usage{data(DUKE.wer)}
\format{
 }
\details{
}
\references{
R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag
}
\examples{
plot(DUKE.wer,type="l")
}
\keyword{datasets}
