\name{DukeLRet}
\alias{BCofLRet}
\alias{DukeLR}
\docType{data}
\title{Duke daily log-returns}
\description{
  Numeric vector giving the daily log-returns
}
\usage{data(DukeLRet)}
\format{
}
\source{
R. A. Carmona: Statistical Analysis of Financial Data in S-Plus, (2004) Springer Verlag, Chapter 2 }
\examples{
data(DukeLRet)
data(EnronfLRet)
plot(DukeLRet,EnronLRet)
}
\keyword{datasets}
