\name{empirical.copula}
\alias{bb1.copula}
\alias{bb2.copula}
\alias{bb3.copula}
\alias{bb4.copula}
\alias{bb5.copula}
\alias{bb6.copula}
\alias{bb7.copula}
\alias{frank.copula}
\alias{galambos.copula}
\alias{gumbel.copula}
\alias{husler.reiss.copula}
\alias{joe.copula}
\alias{kimeldorf.sampson.copula}
\alias{normal.copula}
\alias{normal.mix.copula}
\alias{tawn.copula}


\title{Different copula classes}

\description{
      We provide different copula classes
}
      
\usage{
       
        empirical.copula(U,V)

}

\arguments{ 
	\item{U}{A numerical vector of sample data values}
	\item{V}{A numerical vector of sample data values}
     
}

\value{    
       object of the class empirical.copula
}



\seealso{
\code{\link{fit.copula}}
}


\author{Rene Carmona, \email{rcarmona@princeton.edu}}
\keyword{estimation}
