\name{ENERGY.ts}
\alias{ENERGY.ts}
\docType{data}
\title{Constellation Energy, Natural Gas, S&P500, and Treasuty Bill Readings}
\description{
timeSeries object used in a problem set on Monte Carlo computations from Auto Regressive models with non Gaussian innovations.
}
\format{}
\references{
R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag
}
\examples{
plot(ENERGY.ts[,1])
}
\keyword{datasets}
