\name{EnronIndex.ts}
\alias{EnronIndex.ts}
\docType{data}
\title{
Values of the Enron Index
}
\description{
timeSeries object used in Chapter 8 of the book "Statistical Analysis of Financial Data in R" to illustrate the use of a GARCH(1,1) model with a financial time series.
a problem set on Monte Carlo computations from Auto Regressive models with non Gaussian innovations.
}
\details{The data was generated with a Dow Jones proprietary algorithm. The index is initialized at the value 100 on Jan. 1, 1992. It is computed up until June 28, 2000.}
\references{
R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag
}
\examples{
plot(EnronIndex.ts)
}
\keyword{datasets}
