\name{ethanol}
\alias{ethanol}
\docType{data}

\title{Ethanol data}
\description{
Classical data set used in many text books. Used in the textbook as an example of non-parametric regression in Chapter 4, and as example for the properties of the function \code{lines}.}
\usage{data(ethanol)}
\format{
  A data frame with 88 observations on the following 3 variables.
  \describe{
    \item{\code{NOx}}{NOx concentration}
    \item{\code{C}}{Numeric vector giving the compression ratio}
    \item{\code{E}}{Numeric vector giving the equivalence ratio}
  }
}
\details{
    NOx  C     E
1 3.741 12 0.907
2 2.295 12 0.761
3 1.498 12 1.108
4 2.881 12 1.016
5 0.760 12 1.189
6 3.120  9 1.001
}
\source{
 ????
}
\references{
  ~~ possibly secondary sources and usages ~~
}
\examples{
data(ethanol)
}
\keyword{datasets}
