\name{FRWRD}
\alias{FRWRD}
\docType{data}
\title{Natural gas forward contracts}
\description{
 Values in US Dollar of the 36 Henry Hub natural gas forward contracts traded on March, 23, 1998.
}
\usage{data(FRWRD)}
\format{
  The format is:
 num [1:36] 2.35 2.39 2.40 2.41 2.42 ...
}

\examples{
data(FRWRD)
## maybe str(FRWRD) ; plot(FRWRD) ...
}
\keyword{datasets}
