\name{hills}
\alias{hills}
\docType{data}
\title{Hill races}
\description{
  A data frame of 35 Scottish hill races.
}
\usage{data(hills)}
\format{
  A data frame with 35 observations on the following 3 variables.
  \describe{
    \item{\code{dist}}{a numeric vector}
    \item{\code{climb}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}
\examples{
data(hills)
## maybe str(hills) ; plot(hills) ...
}
\keyword{datasets}
