\name{HOWAREYOU}
\alias{HOWAREYOU}
\docType{data}
\title{Digitized Sound}
\description{
Numerical vector giving sampled values of a recording of the sentence "How are you?".
Used as an example of regularly spaced signal.
}

\references{
R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag, Chapter 6}
\examples{
plot(HOWAREYOU, type="l")
}
\keyword{datasets}
