\name{L1}
\alias{L1}
\alias{L2}
\alias{LOSS1}
\alias{LOSS2}
\docType{data}
\title{
Loss Data Sets
}
\description{
Data needed for the Problem on Operational Risk
}
\usage{data(L1)}
\format{
  The format is:
 num [1:312] 2708 1184 300 533 0 ...
}
\source{
R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
}
\references{
R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
}
\keyword{datasets}
