\name{M.wer}
\alias{M1.wer}
\alias{M2.wer}
\docType{data}
\title{Weekly Excess Returns for a Market Portfoio}
\description{
 Numeric vector giving the weekly excess returns for a market portfolio excess returns constructed from the S&P 500 index. The vectors M1.wer and M2.wer are subsets of M.wer covering disjoint periods discussed in the reference text. 
}
\usage{data(DUKE.wer)}
\format{
 }
\details{
}
\references{
R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
}
\examples{
plot(M.wer,type="l")
}
\keyword{datasets}
