\name{MID1}
\alias{MID1}
\docType{data}
\title{ Data Set used for a Problem on the Comparison of Least Squares and Least Absolute Deviations Polynomial Regression}
\description{
Numeric matrix with 134 rows and 2 columns 
}
\usage{data(MID1)}
\references{
R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag}
\author{Rene Carmona, \email{rcarmona@princeton.edu}}
\examples{
dim(MID1)
# [1] 134   2
}
\keyword{datasets}
