\name{MIND}
\alias{MIND}
\docType{data}
\title{ Data Matrix used in a Problem Set}
\description{
134 x 4 numerical matrix used in a problem of Chapter 5 on Local and NonParametric Regression of "R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag"
}
\references{
 R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag}
\examples{
head(MIND)
}
\keyword{datasets}
