\name{PSPOT}
\alias{PSPOT}
\docType{data}
\title{
Power Spot Price data
}
\description{
Daily Palo Verde firm on peak spot price of electricity between January 4, 1999 and August 19, 2002. Data set used in one of the problems of Chapter 2 of  
"R.A. Carmona, Statistical Analysis of Financial Data in R, 2013, Springer Verlag".

}
\usage{PSPOT}
\format{Numeric vector of length 946}
\details{}
\source{}
\references{
"R.A. Carmona, Statistical Analysis of Financial Data in R, 2013, Springer Verlag".
}
\examples{
plot(PSPOT,type="l")
}
\keyword{datasets}
