\name{ROCK}
\alias{ROCK}
\docType{data}
\title{
ROCK data set
}
\description{
Standard data set used as illustration in a problem of Chapter 5 entitled Local and NonParametric Regression
of " R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag"
}
\usage{data(ROCK)}
\format{
  A data frame with 48 observations on the following 4 variables.
  \describe{
    \item{\code{area}}{a numeric vector}
    \item{\code{peri}}{a numeric vector}
    \item{\code{shape}}{a numeric vector}
    \item{\code{perm}}{a numeric vector}
  }
}
\references{
 R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
}
\examples{
head(ROCK)
## maybe str(ROCK) ; plot(ROCK) ...
}
\keyword{datasets}
