\name{Rsafd-package}
\alias{Rsafd-package}
\alias{Rsafd}
\docType{package}
\title{
Functions and Data for the book Statistical Analysis of Financial Data in S-Plus
}
\description{
Functions and Data for the book Statistical Analysis of Financial Data in S-Plus by Rene A. Carmona (2002)
}
\details{
\tabular{ll}{
Package: \tab Rsafd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-09-15\cr
License: \tab What license is it under?\cr
}
read the book and enjoy the package.
}
\author{
Rene A. Carmona

Maintainer: improved by Paul Hofmarcher, Christoph Leitner
}
\references{
Statistical Analysis of Financial Data in S-Plus, Rene A. Carmona (2002)
}

\keyword{ package }
