\name{SPFUT}
\alias{SPFUT}
\docType{data}
\title{Daily S&P Futures Contracts with Constant Maturities}
\description{
Data set of daily log-returns of two "rolling futures contracts" with constant maturities 3 weeks and 1 week, produced by interpolation. Data set used in one of the problems of Chapter 3 of  
""R.A. Carmona, Statistical Analysis of Financial Data in R, 2013, Springer Verlag"".
}
\usage{SPFUT}
\format{
Numeric matrix with 10483 and 2 columns.}
\references{
R.A. Carmona, Statistical Analysis of Financial Data in R, 2013, Springer Verlag.
}
\examples{
head(SPFUT)}
\keyword{datasets}
