\name{STRENGTH}
\alias{STRENGTH}
\docType{data}
\title{
STRENGTH data set}
\description{
Standard data set used in a problem of Chapter 4, Parametric Regression, of the textbook " R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag"
}
\usage{data(STRENGTH)}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
    \item{\code{strength}}{a numeric vector}
    \item{\code{attenuation}}{a numeric vector}
  }
}
\references{
 R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
 }
\examples{
head(STRENGTH)
## maybe str(STRENGTH) ; plot(STRENGTH) ...
}
\keyword{datasets}
