\name{SUBSP}
\alias{SUBSP}
\docType{data}
\title{
Subset of the daily S&P daily close data
}
\description{
Subset of the daily S&P daily close needed for a problem in Chapter 5, Local and NonParametric Regression
of the textbook "R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag"
}
\usage{data(SUBSP)}
\format{
  A data frame with 1000 observations on the following variable.
  \describe{
    \item{\code{Col1}}{a numeric vector}
  }
}
\references{
 R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
}
\examples{
head(SUBSP)
## maybe str(SUBSP) ; plot(SUBSP) ...
}
\keyword{datasets}
