\name{TC}
\alias{TC}
\docType{data}
\title{
TC data set}
\description{
Temperature and commodity price fluctuations used in a problem set of Chapter 3 of "R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag"
}
\references{
R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
}
\examples{
head(TC)
}
\keyword{datasets}
