\name{us.bis.yield}
\alias{us.bis.yield}
\docType{data}
\title{US yield data}
\description{
  US yield curve provided by the bank of international settlement.
}
\usage{data(us.bis.yield)}
\format{
  A data frame with 1352 observations on the following 11 variables.
  \describe{
    \item{\code{Col2}}{a numeric vector}
    \item{\code{Col3}}{a numeric vector}
    \item{\code{Col4}}{a numeric vector}
    \item{\code{Col5}}{a numeric vector}
    \item{\code{Col6}}{a numeric vector}
    \item{\code{Col7}}{a numeric vector}
    \item{\code{Col8}}{a numeric vector}
    \item{\code{Col9}}{a numeric vector}
    \item{\code{Col10}}{a numeric vector}
    \item{\code{Col11}}{a numeric vector}
    \item{\code{Col12}}{a numeric vector}
  }
}
\details{
     Col2    Col3    Col4    Col5    Col6    Col7    Col8    Col9   Col10   Col11   Col12
1 5.59957 7.22989 7.60820 7.72886 7.75856 7.74944 7.73191 7.71845 7.75699 7.79656 7.80871
2 5.63497 7.17471 7.51762 7.65834 7.69625 7.68582 7.67083 7.66179 7.69745 7.74334 7.75632
3 5.61633 7.20652 7.55039 7.71713 7.75815 7.74095 7.72610 7.71934 7.75459 7.79603 7.80092
4 5.71211 7.20071 7.54165 7.71308 7.76433 7.74835 7.73323 7.72261 7.74459 7.78738 7.79996
5 5.78938 7.19130 7.56918 7.73931 7.78977 7.77982 7.76339 7.73900 7.77044 7.80845 7.81773
6 5.77390 7.14409 7.53450 7.68514 7.73473 7.71907 7.69908 7.68651 7.72154 7.76073 7.77132
}
\source{
  Bank of International Settlements
}
\references{
R. A. Carmona: Statistical Analysis of Financial Data in S-Plus, (2004) Springer Verlag, Chapter 4.
}
\examples{
data(us.bis.yield)
}
\keyword{datasets}
