\name{UTILITIES}
\alias{UTILITIES}
\docType{data}
\title{
Utiities data set
}
\description{
The data set UTILITIES is a matrix with two columns, each row corresponding to a given day.The first column gives the log of the weekly return on an index based on Southern Electric stockdaily close and capitalization, (variable X), and the second column gives, onthe same day, the same quantity for Duke Energy (variable Y ), another largeutility company. Used in a problem set in Chapter 3 of the textbook "R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag".
}
\references{
R.A. Carmona, Statistical Analysis of Financial Data in R, (2013), Springer Verlag
}
\keyword{datasets}
