\name{vw}
\alias{vw}
\docType{data}
\title{VW_Intraday data}
\description{VW Intraday stock values for the period 2008-10-24 -- 2008-10-29.
}
\usage{data(VW_Intraday)}
\format{
 Object of class \code{data.frame}.
}
\details{
       DateGMT   Open   High    Low   Last Volume1 2008-10-24 07:10:00 217.90 219.80 213.71 215.77 2751132 2008-10-24 07:20:00 215.71 215.98 210.83 211.29 2087453 2008-10-24 07:30:00 211.28 215.00 211.22 212.85  846004 2008-10-24 07:40:00 212.50 213.48 209.31 209.61 1459765 2008-10-24 07:50:00 209.62 212.00 209.06 211.13 1278476 2008-10-24 08:00:00 210.70 211.50 210.00 210.46  36825
> }
\references{
http://de.wikipedia.org/wiki/Datei:VW-Intraday-oct2008.png}
\examples{
data(vw)
 head(vw) 
}
\keyword{datasets}
