\name{BASKETBALL}
\alias{BASKETBALL}
\docType{data}
\title{Basketball data}
\description{Basketball data used in regression problems.
}
\usage{data(BASKETBALL)}
\format{
 Object of class \code{data.frame}.
}
\details{
      V1     V2  V3    V4 V5         V6
1 0.5885 0.0888 201 36.02 28   R.Miller
2 0.8291 0.1399 198 39.32 30   M.Jordan
3 0.4974 0.0747 198 38.80 26     D.West
4 0.5772 0.0983 191 40.71 30   J.Dumars
5 0.5703 0.1276 196 38.40 28 M.Richmond
6 0.5835 0.1671 201 34.10 31  C.Drexler
> }
\references{
R. A. Carmona: Statistical Analysis of Financial Data in S-Plus, (2004) Springer Verlag, Chapter 3}
\examples{
data(BASKETBALL)
## maybe str(BASKETBALL) ; plot(BASKETBALL) ...
}
\keyword{datasets}
