\name{BASKETBALL}
\alias{BASKETBALL}
\docType{data}
\title{Basket ball data}
\description{Data used in two regression problems}
\format{
  A data frame with 96 observations on the following 6 variables.
  \describe{
    \item{\code{points}}{a numeric vector}
    \item{\code{assists}}{a numeric vector}
    \item{\code{height}}{a numeric vector}
    \item{\code{minutes}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{name}}{a factor with levels giving the names of the players}
}
}   
\source{R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag, Chapter 4}
\keyword{datasets}
