\name{BCofLRet}
\alias{BCofLRet}
\alias{CCofLRet}
\docType{data}
\title{Coffee log-returns}
\description{
  Numeric vector giving the daily log-returns on the close of the Front Month futures contract of Brazilian coffee 
  (\code{BCofLRet}) and Colombian coffee (\code{CCofLRet}) from January 10, 1986 to January 1, 1999.
}
\usage{data(BCofLRet)}
\format{
}
\source{
R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag, Chapter 3 }
\examples{
plot(BCofLRet,CCofLRet)
}
\keyword{datasets}
