\name{dmvnorm}
\alias{dmvnorm}
\alias{pmvnorm}
\alias{rmvnorm}

\title{Utilities for the Multivariate Normal (Gaussian) Distribution }
\description{
The function {\code{dmvnorm}} computes the multivariate normal density;

The function {\code{pmvnorm}} computes values of the multivariate normal c.d.f.;

The function {\code{rmvnorm}} generates random samples from the multivariate normal distribution.

The mean of the distribution may be a d-dimensional vector. 
The covariance matrix should be given as a d x d non-negative 
definite matrix if supplied with the parameter \code{cov}. 
It can also be given,  if cov is missing, by a vector \code{sd} 
for the marginal standard deviations and a scalar \code{rho} 
implying a constant correlation between all the marginals. 
If the covariance structure of the marginals is supplied in this way,  
\code{sd} should be a d-dimensional vector,
and \code{rho} should be scalar. The dimension \code{d} may be 
inferred from other arguments.
}

\usage{
dmvnorm(x, mean, cov, sd, rho, sigma = cov, log = FALSE)

rmvnorm(n,mean,cov,d,sd,rho,sigma=cov,method=c("svd", "chol"))
}

\arguments{
  \item{n}{ Number of samples generated by \code{rmvnorm}}
   \item{x}{ n x d numeric matrix, each row giving a point at which the density is computed}
  \item{mean}{ d- dimensional vector giving the mean of the distribution }
  \item{cov}{ d x d matrix giving the covariance matrix of the distribution }
  \item{d}{ dimension of the desired random variates (required for rmvnorm)}
  \item{sd}{ d- vector of the marginal standard deviations }
  \item{rho}{ number giving the constant correlation when the covariance matrix is given by its diagonal and the parameter \code{rho} }
  \item{d}{ dimension of the distribution }
  \item{sigma}{ used for compatibility with S-Plus }
  \item{log}{boolean for logarithmic scale }
  \item{method}{String giving the method SVD or Choleski used}
}
\details{
The code of this function is a mere wrapper for the function with 
the same name from the library \code{mvtnorm}. It was written to 
provide compatibility with S-Plus, hence the long list of parameters
}
\value{
  \item{For rmvnorm}{ A \code{n x d} matrix whose \code{d} columns are samples of size \code{n} from the distribution}
  \item{For dmvnorm}{ Vector of length \code{n} giving the values of the density computed at the \code{n} rows of the argument \code{x}}
}
\author{Rene Carmona, \email{rcarmona@princeton.edu}}
\references{
R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag, Chapter 3}
\examples{
TSAMPLE<-rmvnorm(n=128,mean=rep(0,2),sd=rep(1,2),rho=.18) 
TDENS <- kdest(TSAMPLE[,1],TSAMPLE[,2]) 
}
\keyword{multivariate}
\keyword{distribution}
