\name{DSP}
\alias{DSP}
\docType{data}
\title{
Daily S&P 500 
}
\description{
Daily closing values of the S&P 500 index between January 3, 1950 and August 20, 2010.
}

\details{
DSP is the data slot of the \code{timeSeries} object \code{DSP.ts}
}
\source{R. A. Carmona: Statistical Analysis of Financial Data in R, (2013) Springer Verlag
}

\examples{
data(DSP)
head(DSP) ; plot(DSP.ts); plot(DSP.ts)
}
\keyword{datasets}
